0.15: Assert Invalid Tardis Ops in Summary (#2349)
I'm seeing some summaries that have tardis ops below the sequence number of the summary, and the shared string segments. The shared string segments should always be summarizing at the min sequence number, and the tardis ops should bring it forward, so have ops below that is invalid. I haven't been able to figure out the root cause, but this will make the failure more explicit when the issue is hit

related to #2139
Co-authored-by: Tony Murphy <anthonm@microsoft.com>