yo-fluid updates (#2323)
This change is pretty large but it's a lot of refactoring. The main changes are:

Streamline the developer flow
Now only 2 inputs: ComponentName & View type
optional cmd line args
Splitting View/Component/Model into separate files
Lots of Documentation in both the generator README and the output README
Reducing the package.json scripts to a smaller set
Remove unnecessary root files
Remove container option (we might want to add this back later but for now we are streamlining). Adding container back later also becomes easy because we now have index.ts as an explicit export file.
Update packages to make sure the output works for now.
Notes
The documentation isn't 100% yet because we still haven't converged on the yo package name.
You might also notice that while we have a .npmrc I don't talk about it because it will go away.

Things that are not covered but still need to happen
Add testing to ensure when we do version bumps we have a way to test the output of the generator.
upgrade the packages and package names to ^0.19.0