Add end-to-end tests to yo-fluid (#2378)
This change add jest-puppeteer tests to the output of the yo-fluid generator.

It also adds two tests to the yo-generator itself that uses those above tests to validate both the react and none views work end to end.

This includes generating the temp file, running npm i running npm test and validating the output.

Note: this adds about 3-4 minutes to the CI loop for the yo-generator. This is because it does an npm install and npm test (starts the server navigates and such).

I'm not concerned about the extra time since the yo generator is running in its own CI loop.