Strongly Type Group Ops To Prevent Nesting (#2408)
As part of the reconnection work, I also introduced this change to increase type safety. Breaking it our in a separate change to reduce the size of that PR.

This change limits the possible ops with-in a group op to non-group ops in order to prevent nesting of group ops in group ops.