Rename MergeTree Snapshot Blob to CatchUp Ops (#2372)
Adds a contains method to ObjectStorageService so we can determine if a blob exists in the tree. Add support for a new name for catchupOps. Currently we still emit with the old name, so that old clients won't break. Once this change is consumed, in the next version we can emit snapshot with the new name by default. If we need to completely get rid of the old name, we will need to help from sharepoint to modify files at rest.
Co-authored-by: Tony Murphy <anthonm@microsoft.com>