ReSubmit Sequence Ops in Order on Reconnection (#2391)
* ReSubmit Sequence Ops in Order on Reconnection

* handle group ops and annotate

* fix replay toolThis change integrates the recent changes to have the runtime dispatch ops on reconnect, so they are replayed in order. With this change, sequence DDS will replay their ops in order.

For every operation applied to a merge tree, which is the underlying DDS for sequence, we create a segmentGroup. A segmentGroup is a collection of segments, which is bi-directionally linked to those segments, so if a segment is split or merged, the segments in the segment group are also split or merged. These segment groups are stored in a queue, which is popped on ack of the op, so the segments can be update with the sequence information.

We also used these segment groups for the reconnect case, as from the segment group we can rebuild the necessary operations that would result in the segments represented by the segment group.

The major change here is the addition of the localSeq/localRemovedSeq to segment groups, and segments. Previously we would re-gen all pending ops together, so we didn't need to differentiate them, but now that we want to generate them one by one and in order, we need to only take into account operations that would be applied by the time the regenned op is applied. Every segment group now has a localseq, so when regenning the ops, we can exclude ops with a localseq greater than that of the segment group.

Related #2382 Fixes #1780

* get pending segment groups in correct order

* track localSeq in merge tree

* add some comments and logging

* fix typo

* add multiple iterations to test

* pr feedback

* add findReconnectionPostition to client

* Add stress tests for generateing reconnect messages

* fix lints

* fix change to reconnectFarm rounds

* fix bad merge

Co-authored-by: Tony Murphy <anthonm@microsoft.com>