Sequence: Remove Old Catchup Op Blob Name (#2470)
This change fully removes the old catchup blob name from the this repo, and the test data repo. In addition i've added a new mode to the replay tool and the snapshot replay tests that loads a snapshot from the extended format or storage format, and replays ops on top of it. This allows us to maintain compatibility with legacy formats. The collateral for these is storage in the snapshot test content folder under original_snapshots.

* Remove Old Catchup Op Blob Name

* remove instance in readme

* pr fixes

* support loading old snapshots in snapshot tests

* tsfmt

* use FileSnapshotReader

* debug logs

* remove debug code

* increase error reporting

* fix error message

* tsfmt

* fix packageVersion regex

* tsfmt

* match dot over digits

Co-authored-by: Tony Murphy <anthonm@microsoft.com>