MessageType: break into 3 distinct sets (loader, container, component): phase 1 (#2496)
The purpose of this change is to
1. Ensure there is proper wrapping / unwrapping of ops across layers, such that specifics of each layer do not leak to next layer
2. Ensure that op types do not overlap between layers, i.e. each uses distinct types and they do not form one global list.

Notes:
1. I did not remove ability of ContainerRuntime to specify MessageType, it's hard due to RemoteHelp & Summarize ops, which are not runtime ops.
In future changes, I'll add assert on Loader level to assert that runtime can send anly 3 types of ops - op, remote help & summarise. In future, we can break it into 3 APIs and remove type from API at all.
2. I'm keeping names ("op") for back compat, but changing internal names to better reflect their meaning (i.e. ComponentMessageType.ChannelOp).

Maintaining back compatibility, will switch to "wrapped" mode in later version of runtime

Moving code around to unify handling of ops. I.e. layer that wraps content with IEnvelope is also the layer that unpacks it.

Deprecated:
1. Ability for components to force snapshotting (save() method)
2. MessageType.BlobUploaded - nobody uses it. We will design it once we need it.

There are some small unrelated changes in preparation to aliasing changes (as this change was born from that work).