General OSS Cleanup (#2568)
There are a few functional things in here:

Add a comment to component-base that it is experimental. It's used to widely to push into anything specifically.

merge server-tools-core with scribe because it was the only component using it.

move local-test-utils from ./packages/server/ to ./packages/test/ There was already a utils there called local-utils so I renamed the folders to match the packages. We probably want to consolidate but it's out of scope from this change.