Support split view/model in Spaces (and Waterpark) (#2570)
With this change, SpacesStorage supports storing arbitrarily complex serializable data. SpacesStorageView correspondingly supports referring to an external authority (ultimately spacesItemMap currently) on how to extract a view from that data, rather than assuming the data is a single handle to a ReactViewAdapter-compatible component. As a result, we can support split view/model in Spaces, since that authority is free to join models and views however it likes.

To demonstrate this capability, I've added usage of the multiview Coordinate and SliderCoordinateView sample's split view/model.