Lift Sheetlet from table-document to table-view (Fixes #2699) (#2700)
This is the smallest change possible to remove the sheetlet code from table-document.

I have not changed the ITable api and removed evaluateFormula and evaluateCell - they now just return default values. (I don't know if anyone was using them). I'm happy to remove the functions or throw an error, but someone will need to track down the consumers that might be broken by changing ITable.

I didn't change table-document to implement IMatrix. I didn't think it was worth it because I'm not sure how long table-document will be around. Again, I'm happy to change it if needs be.

Fixes #2699