Support Partial Load on Sequences (#2689)
This change enabled partial load of sequences when the configuration option sequenceInitializeFromHeaderOnly is set to true.

When this is set to true, we load the header, and then return from the load core call after the header is loaded. While the body is loading we cache all incoming an outgoing ops. Once the body is loaded we apply the incoming ops, and then resubmit the outgoing ops.

Fixes #89