Host and Loader Contracts Forward Compatibility (#2769)
This change proposes a pattern for component interfaces that may cross the loader boundary. For these interface we will create a new copy with the new names, but preserve the old interfaces. We will not consume these immediately, but ship them in a release to our partners. They will need to consume these changes in their hosting layer, and use the nullish coalescing operator (??) between the old and new interface. Due to duck typing, this should allow using the returned type as normal.

In a later release we will remove the legacy interfaces for our code base. If partners still need them they can keep copies in there code base, or cast to any, to keep support for backwards compatibility.

related #2758
related #2764