Add headers telemetry property for fetch() calls; SPO code cleanup (#2835)
Add headers:true property for any fetch() call we do if headers were used. That's good indication on where OPTIONS calls is made (unless it's cached by browser), which doubles number of network calls. This is useful in telemetry, especially in places where we dynamically switch to using headers because URI with token is too long.

Closes #2557: createNewFluidFile use of cache seems wrong

Properly documents response type from SPO for different ways we create files: ICreateFileResponseZeroFile vs. ICreateFileResponse

Removing IFileCreateResponse altogether because it's impossible to return such structure. Creation file from snapshot does not return name of file, so we do not know it. Returning requested file name (which is renamed on collision) is incorrect and may lead to bugs.