Disposed unhandled exceptions in CRC / Agent scheduler (#2837)
Closes #2734: QoS Report Unhandled Rejection: ComponentRuntime disposed while this ack-based Promise was pending

Writing flows are mostly fire-and-forget in CRC.
There is not much value in returning failure on closure of container, as it does not matter.
The only place where it matters (and API shape does not allow to swallow error) is waitAndAcquire() API, where expectation is clear that when it returns, it has to acquire an item.

The caller can't do much about these failures anyway, so there is not much benefit of propagating to users.

We also should consider other design. We can remove all pending op tracking (i.e. clear unacked ops) on disposal (it's good thing to do either way!) and simply rely on GC to cleanup any hanging promises, without need to rely on newAckBasedPromise. I feel it's cleaner design, as we keep adding catches for runtime disposed exceptions here and there without much benefit.
Issue #2852 is opened to track this work.