Rename of _createComponentWithProps to _createComponent
IContainerRuntimeBase.getComponentRuntime() is gone, instead there is getComponentById that returns IComponent, not channel. It automatically does a request for component after finding component runtime for given ID.
I have a feeling that right next step here is to remove this API and base into ContainerRuntime.request() optional behaviour of returning components by ID, as all the code that uses this API is doing just that - it's all default request handlers.
Future step would be to limit it only to named singleton components.

Back compat tests fail with this change - need to fill in back compat story.