Properly handling readonly permission transitions (#2908)
Currently, ContainerRuntime will send ops if user has read-only permission to a file, This ops would be rejected by DeltaManager, but result is not check anywhere.

First step is ensure that lower layers do not receive anything back (void) as result of sending op. They used to get client Seq number, but now container runtime manages all resend logic, so for all lower layers ack of submitting op is fire-and-forget.

At deltaManager level we are moving to close container if op is sent in read-only permission state

At container runtime level, we handle changes between states better, and do not send ops, but rather accumulate them. Ideally there should be no ops, as all components should react to read-only state and not send anything. But due to asynchrony / offline, this is not possible in all cases. An issue is opened to track warning container about such cases such that host can react in some way to such condition.
If user gets write permissions to a file, all pending ops will be flushed to server.

This change also adds requirements around when Container.forceReadonly() can be called, to ensure that we resend ops correctly, and do not attempt to send same ops twice. Review of API is required (as a follow-up).