Add Node Buffer abstraction in common-util (#2887)
Part 1 of #2450
Add a minimal Node Buffer abstraction in common-util that provides the functionality that we currently use.
Add a jest test for the browser implementation. This exists separately from the mocha tests because of typing issues between Node and browser (e.g. tsc gets confused about the return type of setTimeout, which differs depending on the environment).

Followups:
Remove sha.js usage, which couldn't have been done before because of old Edge support (and which has a dependency on buffer)
Update consumers