update buffer usage in common utils
copy the contents of base64encoding into each of bufferNode/bufferBrowser.  we do it this way because we only tell webpack to look at the indexNode/indexBrowser files (which each reference the _node/_browser sources). keeping the same copy of base64encoding would mean we need to add additional files to watch in the package.json, where keeping the split at the index level is simpler.  the base64 conversion fns are small so this is relatively cheap in terms of duplication, but if larger amounts of code run into this we can revisit how things get split between node and web