Copy over bundle buddy from bohemia (#3211)
Copying over (most) of bundle buddy from bohemia to make the upcoming diffs more readable, with only minor changes to satisfy the repo policy check and remove references to ffx/bohemia.  None of this is building as part of any CI yet.  I left out some of the files that would end up being deleted anyway.

Upcoming changes are primarily refactoring to make this more reusable, and followup from that would be for bohemia to consume the reusable bits through fluidframework (probably with bundle buddy existing as the wraparound).

Thanks to christiango and aboktor for writing the original code