add uint8arraytostring helper (#3257)
#2450

Add a helper to convert uint8arrays to strings to work around a buffer dependency in protocol-definitions. ISummaryBlob.content in protocol-definitions had type string | Buffer, but the addition of the isomorphic Buffer implementation IsoBuffer does not allow a direct swap due to layering (the implementation is in common-utils). To keep the definitions vanilla-ly typed, swap Buffer here for Uint8Array and add a helper to do the equivalent Buffer.toString conversion.