Add Live Fluid Dice-Roller in Docs (#3355)
This change adds a side-by-side dice-roller example to the docs. It adds it to both the QuickStart and Tutorial page.

It works by using session-storage and always generates a new document on load. The bundle is uploaded to the cdn and it uses the standard side-by-side html template.

I've added a shortcode so that we can re-use this with other examples through the docs. I tried getting the three examples working but because the space is really limited (with side-by-side plus in a tiny width div) the solutions need to be a bit custom and will take a bit more time.