add # search builds option
add option to change defualt number of builds to search. potentially useful for us because we have a high number of builds (especially if you haven't merged master very recently).  also change it to only pull builds once so we don't hit the ado api a bunch when doing fallback