Track additional clientIds from previous connections (#3613)
Fixes #3605
The suspicion is there is a timing issue around reconnects where we end up with multiple connections with pending ops that can be linked back to the original connection, but not each other resulting in multiple submissions of the same ops. We haven't identified where exactly that is in the code, so we want to add additional telemetry and additional clientId handling to allow us to tell that both connections in this case are the same client.

Keep track of the previous ~~5~~ 💯 clientIds in the delta manager instead of just the last one.