Minimally restore _createDataStoreWithProps for bohemia (#3635)
#3631
Minimally restore _createDataStoreWithProps. This includes adding back the API and adding back passing initial configs through the context. Bohemia isn't quite ready to move off this API yet but they will be soon™.

Future: port this change to 0.26 and main branches as well (Bohemia is not expected to be ready by 0.27 release)