Add level back to tinylicious
Being able to store ops on disk, rather than fully in memory, makes tinylicious scale better as a service. The functionality used to exist but was removed due to an NPM package security warning. A prior version of level-sublevel does not take a dependency on the package with the security warning.