Throw on errors in replay tool instead of logging to console (#3735)
Fixes #3107

Currently the telemetry logger used in the replay tool will print error details (such as from a container error) as console errors. The snapshot tests do not have any specific assertions and instead rely on promise resolution to pass/fail, so this functionally hides swallows failures and causes failing tests to report succeeded. Change the logger to re-throw the error in these cases, causing tests to fail as expected.

Potentially all console errors should be made to fail tests, but that's a different discussion.