Remove Shared Object FluidObject Interfaces (#3782)
Remove the unused interfaces on FluidObject

ISharedString
ISharedObject
IChannel
In addition to being unused these interfaces represent bad practices. A DataStore shouldn't externally expose it's dds for others to access directly. It should instead expose and API and control the data model of the dds.

While gernerally unused there is a coupling between our scribe tool and our shared text examples that leverages the ISharedString interface. To remove this interface from IFluidObject, which broadly exposes this bad practice, I've move it to using the request pattern. While still not ideal, this localizes the contract between the two rather than leaking this across the system.

Related #3264
Related #2566