Deprecate submitAsync (#3818)
submitAsync was implemented initially to allow larger message content payload to be uploaded/downloaded asynchronously from the op message itself. Now that we have blob API, that should serve the same purpose.

The functionality isn't being used for a long time now, and shouldn't have back-compat concern.
Clean up most of the client code and also the server support. Will clean up the IConnected.initialContents in the next PR once the server changes in this PR are committed.

Fix #2039