Add script in build-tools for bundle analysis + danger fixes (#3828)
Add a script to run bundle analysis so it doesn't need to be handled in the yaml build step. This is preferable because it better encapsulates its usage and dependencies (using the build step required adding danger and bundle-size-tools dependencies in the root package.json, but also hid how they were being used).
Also some fixes to the danger file and dependency.