Add separate yaml for bundle size artifacts (#3841)
Currently the artifacts are produced as part the regular client CI, but that CI is run in the private ADO project and need to be consumed from the public ADO project. Instead of providing access to the private project (which would overscope the actual permissions we want to provide), move the artifact generation part into the public project.