update search distance (#3882)
There are two variables here for bundle analysis: our commit search distance and our build search distance. The commit search distance is the default number of commits to look back from the common branching commit in main. We need it because not all commits' associated pipeline builds produce size artifacts, and it needs to be sufficiently larger than the average number of consecutive commits not producing size artifacts. The build search distance is the default number of most recent builds to pull from the CI pipeline. We can't pull just a build for a specific commit, so this needs to be sufficiently larger than the average number of commits to main past the branch's common branching commit (i.e. the faster commits go into main the larger this will need to be).

The default number of builds was 20 which was fine, but the default commit distance was 5 which was not large enough. Just use 20 for both by default (the overlap will be smaller); consumers can still adjust these as needed.