Speculative mitigation for inconsistent summary sequenceNumber (#3907) (#3909)
We're seeing instances where the summary claims a .protocol attributes sequenceNumber (e.g. 325) but actually only includes ops up through a prior number (e.g. 322). As a result, new clients loading the document will start reading ops at 326 based on the attributes and hit errors and/or corrupt the document due to the missing data from 323-325.

When this happens, we see the summary op has the data up to 322, a "message" field up to 322 (e.g. "Summary @322:318"), but a referenceSequenceNumber of 325. We protect against the sequence number moving during summarization with an assert already (as of 0.26), but it seems possible the sequence number could move during summary upload without detection. This change adds another assert that the sequence number also does not change during summary upload but before the summary op gets created.