Cleanup of serializer interfaces. (#3886)
Make context be part of serialized object, instead of continuously passing it (and searching for root) on every call.

This also helps future work to disambiguate meaning of IFluidHandleContext at data store level, which can be used to represent different routes (dds, object) - subject of future PRs.

Unrelated, but we may not even need context, other than root context, as all out paths are absolute paths. That said I do not understand yet all implications of back compat (there is comment in there about it) to rip it.