Move Loader Constructor to Props Pattern (#3945)
As part of some prototyping related to the version work, I saw the need to add another parameter to the loader constructor. I'm not sure what will come to fruition, as I'm continuing to explore a few designs. However, in looking at adding a new param i saw that the current constructor already takes several optional params, which is cumbersome, as each preceding optional param must be specified to specify the next. Moving the loader to a props pattern will make the loader easier to work with and construct.

Additionally, many of these same parameters are then passed to the container. I've simplified this to just pull the properties it needs directly from the load itself.