Restructure ODSP connection telemetry (#3965)
This change restructures the connection to unify connection logic and telemetry on AFD vs. non-AFD. The distinction of cache vs. first-attempt vs. fallback should be derivable from telemetry without specially-named events. This should also make it easier to observe connection timing (some of the flows don't log enough data to determine this currently).

This change does include one functional change beyond telemetry - currently we only write the AFD-valid cache if there wasn't a valid cache. With this change, we'll write an updated timestamp to the cache on successful AFD connection regardless of the prior state of the cache.