Split debug logger into a different module (#4078)
Our bundle has multiple copies of debug in it. Some of it comes from Fluid directly, others through Fluid dependencies like socket.io.

Unfortunately, the debug module is not tree shake eligible, so if it appears as an import in a module where any of the exports are used, it will be used in the module. To work around this limitation, we should put the usage of debug in it's own module, that way if the entire module is not used, it will be tree shaken.

Debug is still heavily used in the Fluid repo, so more work will have to be done to fully solve this issue.