Add loggers into IFrame host (#4106)
Fixes #1421

Two goals here:

    Provide a way for the inner-IFrame driver to get the outer driver to log messages
    Provide a way for the inner-IFrame driver to log messages itself

The ICombinedDriver proxy object is currently functioning as the driver's internal props, so I added a minimal (ITelemetryBaseLogger) logger in there. This lets the inner-IFrame driver to call send on that logger instead of needing to create its own. Also update the IDocumentService and IDocumentServiceFactory implementations to more closely match the live ones wrt logger arguments, and add an example where the inner-IFrame driver creates and uses its own ITelemetryBaseLogger instead of using the proxy logger.