Fix parsing nested required dependencies in package lock file (#4131)
This tools that generate a package-lock.json file from lerna-lock-file.json so that the component detection tools can find our dependencies. The bug is in the code that mark package and its transitive closure of package dependencies as non-dev packages, The "requires" dependencies in package-lock.json can be resolve from its parent's dependencies set. So keep track of it's nested scopes and check all of them before going to the top level.