Support Skipping Context Reload when New Code is Compatible with Old (#4122)
Adds IFluidCodeDetailsComparer interface that supports comparison, and constraint satisfaction. The interface has two methods: compare and satisfies. Compare can be used to determine if code should be proposed, as it provides relative ordering for the code details. Satisfies determine if a code details satisfies the constraints of another code details. This is used to determine if we can continue using the existing runtime without reload when a new code proposal comes in.

The IFluidCodeDetailsComparer is supported on both the code loader, and the fluid export. The code loader is the host provided entry point and the fluid export is the container authors entry point. The interface is optional on both.

Expose the current code details off IRuntime. This will allow the runtime to reject incompatible changes

Last but not least are a bunch of test updates, and a new test scenario to cover the usages of IFluidCodeDetailsComparer. Specifically, proposing a new code proposal that is satisfied by the existing code proposal, so no reload of the context is necessary.

related #3481
related #3529