use new bundle-size-tools (#4179)
Part 2 of #4124

Consume the new version of bundle-size-tools in build-tools to only post a comment when a size change is detected. In cases where the size does not change, we do not post a comment and danger will delete any old comments. The implication for users is then that if there is no comment, the bundle size is not affected by the pr