Summary fix (#4392)
This should fix #4377

ScheduleManager has multiple way to tracking pause/resume state for patched op. Summary also use it pause the delta inbound during summarization. Since we switch to counting pause/resume, the ScheduleManager could cause the pause/resume out of balance with this sequence:

1. Summary start which calls ScheduleManager.pause() -> (this.pause === true) -> inbound.pause()
2. op got pushed to the inbound queue -> ScheduleManager.updatePauseState -> ScheduleManager.localPause = false, but instead we call pause again because this.pause === true)
3. Summary stops -> ScheduleManager.resume() is called, and inbound.resume() is called
The middle step is what cause the extra inbound.pause() to be called.

The fix is to separate the summarizer pause/resume call and the schedule manager's tracking of batch ops, and have the summarizer directly call deltamanager's inbound queue for pause/resume