force ansi color (#4279)
Build logs in Azure DevOps are black and white instead of the colorized output we see locally. This is because our color library auto-detects color support, and the ADO pipeline environment doesn't work as expected there. We can set an environment variable to force ANSI colors, which ADO pipelines will show correctly.