Encapsulate Contexts Deferred (#4448)
Decided it was best to break apart #4420 as i don't want to introduce unintentional errors whille merging. The end goal is still the same:

ContainerRuntime is overloaded. This PR splits out all the data store handling into it own class. There should be not functional changes here, simply moving existing functionality. Once this is out, we can work to clean up the interface. Eventually datastores could just a partitioned and virtualized DSS, and data stores could host data stores. This could be used to solve issues we see with the flat list around garbage collection, and also fulfill the dream where everything is a channel, as we could compose different container runtimes that are just a different type of dds, rather than data stores at the root. Also, hoping this will increase testabiity of the datastore code, plan to explore that in a follow up.

This PR start's that journey by encapsulating contexts deferred. As as pull in layers: requesting, creating, attaching, and summarizing we can make previous layers like this, private.