Fix op telemetry (#4315)
Issue #3895 - Add telemetry tracking ack of op speed while on active connection
While this change is not 100% what is described in issue, it's not very clear if we need actual implementation it asks for.
At the same time, implementation we have is incorrect and data can't be trusted, so I see fixing it as a first step on re-assessing the needs in this space.

Specially we were resetting clientSequenceNumberForLatencyStatistics on "connected", but ops may come in between disconnect and connected events, because we do fetch in parallel to establishing connection in some scenarios. As result, we are measuring offline time occasionally, which pollutes data.

Also changing naming of certain fields to be more in line with names used elsewhere.