Encapsulate Data Store Methods Not Related to Summarization (#4477)
Decided it was best to break apart #4420 as i don't want to introduce unintentional errors while merging. The end goal is still the same, fully encapsulate data stores with eventual goal of hosting on a channel.

Related #4448 , #4460

This PR encapsulates all remaining methods that are not related to summarization. After we encapsulate summarization, we make contexts private on data stores, and remove the dependecy of the the container runtime on data store contexts