Improve ODSP socket reuse logic (#4483)
We have some number of errors in telemetry that are hard to track down.
One of them is hitting this assert:

assert(key !== undefined, "reentrancy not supported!");

I was not able to fully pinpoint the problem. But couple times I was able to hit it, we had two "disconnect" handlers - one from DeltaManager, one - connection listener (later should have been removed on successful connection). I've put more telemetry logging, and made code safe in those cases, will look for more telemetry to better understand the problem.

Another one is related to raising "disconnect" event on socket as result of closing one connection and causing failure / closure of other connection reusing same socket.
Examining code reveals that decisions on whether socket is alive is wrong. We use disconnected property on the socket as a way to determine if socket can be reused. However, socket might have never had a chance to connect in the first place. In such case, not only we will not reuse it, but we will fail connection that is reusing same socket. This is being addressed by adding isPendingInitialConnection flag and tracking successful usage of the socket (or not).

Fundamentally, the other problem here is having global map and not really well isolated lifetime management of socket references. Using keys asks for trouble, as it leaves us in position where cleanup of old connections has to happen (synchronously) before new connection is going to reuse same slot (same key). This is not good, as we will keep running into interesting cases where this requirement is violated.
As such, I'm moving away from using keys in the code, to using actual SocketReference object, and only using key when creating / finding existing SocketReference. Once SocketReference is in bad state, it's removed from the list, but existing users of SocketReference can hold it as long as they need to do their cleanup. New users of the system will create new SocketReference but it will not interfere with old users who are going through cleanup.
This also allows us to reduce global state sharing across users of socket. In particular, we raise "disconnect" event on socket only in response to "server_disconnect" event and no longer do it for any other critical events on the socket, with expectation that all users will process individually whatever event causes us to believe socket is bad.
And with this, logic moved to SocketReference - place where it should have been in the first place, out of OdspDocumentDeltaConnection.