Remove SHA Computation from Flatten Blob (#4536)
This is related to #4310 as to validate data store attributes and fail fast we need synchronous access to the attributes blob. The only thing blocking that is the sha calculation in flattenCore.

It seems safe to replace the SHA with a uuid, as the SHA is only used as an in memory id when creating the snapshotTree. Once buildSnapshotTree is synchronous, it allows the unwinding of a bunch of async code that existed simply to all the async SHA calculation.