Bump common-utils version (#4553)
common-utils got some changes that need to be consumed in server, and where the dependent version needs to match between client and server. Bump the common-utils version to the prerelease version in server

Next steps: publish prerelease version of server (1006.0-0) and bump the consumed versions of common-utils and server in client so that only a single version of common-utils gets used.