Hook our azure dev ops pipeline to our azure key vault (#4644)
In this change we take a dependency on the existence of a variable group, prague-key-vault. In our internal environment we hook this up to pull from key vault to pull real secrets for running real service tests. In our public environments this variable group must exist, but can be empty