(split from Whiteboard) Merged PR 26987: fix: Correct shared-tree lint configuration to not pick up Whiteboard settings
Initial conversion of shared-tree to use fluid's lint rules had an oversight: because of the `.eslintrc` at the root of the repository, we ended up inheriting *both* sets of linting rules. This fixes that problem by using the `root` option in `shared-tree`'s `.eslintrc` and removing disabled overrides as appropriate.

Related work items: #49299