(split from Whiteboard) Merged PR 26984: improvement: invalidate the dependents of all nodes within deleted document subtrees
## Changes

This PR changes the invalidation policy for document subtrees of deleted nodes:
- Before this change, only the projections of the parent and trait that are modified by a delete operation were invalidated.
- After this change, the projections of the parent, trait, and the projections of all the nodes/traits in the deleted subtree are invalidated.

This change has been made to the following `DocumentStore` implementations:
- `ItemTreeDocumentStore` (on which `NaiveDocumentStore` is based)
- `FluidDocumentStore`

This change has no visible impact (aside from performance implications) for the customers of `DocumentStore`.

## Performance Impact

Running the `perf:measure` script shows the performance impact is negligible for either store as variations in timings for either revision are greater than the difference across versions.

## Motivation

The previous policy will become invalid once we allow a node to be replaced with a node whose ID is the same but whose data is different. Under the previous policy the projection for the parent of this node would re-run but would then produce a proxy that passes the sufficiency check when compared to the old proxy (for the original node that had the same ID but different data).